requirements.txt
streamlit
requests
python-dotenv
google-generativeai
thefuzz
python-Levenshtein
sentence-transformers
scikit-learn
pandas
altair
plotly
streamlit-plotly-events
